+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - This beautiful change log
 - Method for mandate revocation
 - Method for updating customer details
+- startDate option for subscriptions
 
 ### Changed
 - Allow metadata to be entered as array or object instead of just array
@@ -14,6 +15,7 @@
 - **Removed** methodParams option in Customer\\PaymentResource $opts parameter. Add method parameters to $opts directly.
 - **Removed** method and webhookUrl parameters from Customer\\SubscriptionResource. Add them to new parameter $opts instead for consistency.
 - **Renamed** $opts parameters bic to consumerBic and reference to mandateReference in Customer\\MandateResource
+- Restructured and improved unit tests
 
 ### Fixed
 - Clean up code and fix most of the open CodeClimate issues
@@ -30,6 +32,6 @@
 ## 1.0.0 - 2016-08-23
 Initial release
 
-[1.1.0]: https://github.com/Cloudstek/mollie-php-api/compare/v1.0.2...v1.1.0
+[2.0.0]: https://github.com/Cloudstek/mollie-php-api/compare/v1.0.2...v2.0.0
 [1.0.2]: https://github.com/Cloudstek/mollie-php-api/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/Cloudstek/mollie-php-api/compare/v1.0.0...v1.0.1